2006-06-23  Matthew Allum  <mallum@openedhand.com>
        * examples/video-player.c: (size_change):
        cast stage height to gint avoiding wierdness for when
        video height greater than half display height. Fixed #99

+++ examples/video-player.c
@@ -200,7 +200,7 @@ size_change (ClutterTexture *texture,
   clutter_texture_get_base_size (texture, &vid_width, &vid_height);
 
   new_height = ( vid_height * CLUTTER_STAGE_WIDTH() ) / vid_width;
-  new_y      = ( CLUTTER_STAGE_HEIGHT() - new_height) / 2;
+  new_y      = ( (gint)CLUTTER_STAGE_HEIGHT() - new_height) / 2;
 
   clutter_actor_set_position (CLUTTER_ACTOR (texture), 0, new_y);
